Make kill command conditional

+++ Jenkinsfile
@@ -87,7 +87,9 @@ pipeline {
         stage('Test MODAK') {
             steps {
                 sh  """ #!/bin/bash
-                docker kill \$(docker ps | grep modak | awk '{print \$1}')
+                if [ -n "\$(docker ps | grep modak)" ]; then
+                    docker kill \$(docker ps | grep modak | awk '{print \$1}')
+                fi
                 docker-compose build --no-cache
                 docker-compose up -d
                 docker exec -it applicationoptimisation_restapi_1 /bin/bash -c "cd ../test; python3 -m unittest -v"
@@ -209,7 +211,7 @@ pipeline {
                     sh """#!/bin/bash
                         hostname
                         ls
-                        ssh -i ${modak_ssh_key_file} -o StrictHostKeyChecking=no ${modak_ssh_username}@192.168.2.155 "cd application-optimization && docker-compose down && cd .. && rm -rf application-optimization; mkdir -p application-optimization"
+                        ssh -i ${modak_ssh_key_file} -o StrictHostKeyChecking=no ${modak_ssh_username}@192.168.2.155 "cd application-optimization && docker-compose down && cd .. && rm -rf application-optimization; docker kill \$(docker ps | grep modak | awk '{print \$1}'); mkdir -p application-optimization"
                         scp -i ${modak_ssh_key_file} -r ./* ${modak_ssh_username}@192.168.2.155:application-optimization/
                         ssh -i ${modak_ssh_key_file} -o StrictHostKeyChecking=no ${modak_ssh_username}@192.168.2.155 "cd application-optimization && docker-compose up -d"
                        """
